Allow derived types in Initialize
When deriving from a class that calls Initialize in it's constructor
(like Solution), the initialization does not occur because the derived
type does not match the type that calls Initialize.  However, if that
derived class then relies on being initialized later in the constructor,
an exception is thrown.

Instead, permit derived types to initialize in the base constructor (in
the use case this is for, we actually want the derived type in the add-
in to behave as the base type for almost every scenario).