Make OnRunTarget public for Project class
As per the "unseal Solution" pull request, we need to redirect
operations performed on Protobuild objects (modules / definitions) to
the underlying MSBuild solutions and projects.  Although we can override
OnRunTarget in the Protobuild classes, we can't actually call
OnRunTarget on the MSBuild classes (that are being proxied / shadowed)
because it's internal protected.  In order to pass the OnRunTarget
operation through to the actual project, we need this method to be
public.